More settings redesign.  This time, separate the hosts lists from the individual sections so we can give it a better place on the page to live.  Also cleans up a bunch of the html/form fields used by the settings and keybindings editors.
git-svn-id: http://svn.mythtv.org/svn/trunk@12284 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ modules/mythtv/set_settings.php
@@ -13,21 +13,29 @@
  *
 /**/
 
-// Specific host?
-    isset($_GET['host']) or $_GET['host'] = $_GET['host'];
-    if (empty($_GET['host']))
-        $_GET['host'] = null;
+// Load all of the known mythtv hosts
+    $Settings_Hosts = array('' => '- '.t('All Hosts').' -');
+    $sh = $db->query('SELECT DISTINCT hostname FROM settings ORDER BY hostname');
+    while (list($host) = $sh->fetch_row()) {
+        if (empty($host))
+            continue;
+        $Settings_Hosts[$host] = $host;
+    }
+    $sh->finish();
 
+// Make sure we have a valid host selected
+    if (!isset($Settings_Hosts[$_SESSION['settings']['host']]))
+        $_SESSION['settings']['host'] = reset($Settings_Hosts);
 // Save?
-    if ($_POST['save']) {
+    elseif ($_POST['save'] && isset($_POST['host'])) {
         foreach ($_POST['settings'] as $value => $data) {
-            setting($value, $_GET['host'], $data);
+            setting($value, _or($_POST['host'], null), $data);
         }
         if (is_array($_POST['delete'])) {
             foreach ($_POST['delete'] as $value => $data) {
                 if (!$data)
                     continue;
-                if (is_null($_GET['host']))
+                if (empty($_POST['host']))
                     $sh = $db->query('DELETE FROM settings
                                        WHERE value=? AND hostname IS NULL',
                                      $value
@@ -36,25 +44,17 @@
                     $sh = $db->query('DELETE FROM settings
                                        WHERE value=? AND hostname=?',
                                      $value,
-                                     $_GET['host']
+                                     $_POST['host']
                                     );
             }
         }
+    // Make sure the session host gets updated to the posted one.
+        $_SESSION['settings']['host'] = $_POST['host'];
     }
-
-// Load all of the known mythtv hosts
-    $Hosts = array();
-    $sh = $db->query('SELECT DISTINCT hostname FROM settings ORDER BY hostname');
-    while (list($host) = $sh->fetch_row()) {
-        if (empty($host))
-            continue;
-        $Hosts[] = $host;
-    }
-    $sh->finish();
 
 // Load all of the settings for the requested host
     $MythSettings = array();
-    if (is_null($_GET['host']))
+    if (empty($_SESSION['settings']['host']))
         $sh = $db->query('SELECT value, data
                             FROM settings
                            WHERE hostname IS NULL
@@ -64,7 +64,7 @@
                             FROM settings
                            WHERE hostname=?
                         ORDER BY value',
-                         $_GET['host']);
+                         $_SESSION['settings']['host']);
     while (list($value, $data) = $sh->fetch_row()) {
         $MythSettings[$value] = $data;
     }
